pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Compare 2 json or yaml files"
authors = ["ImTsukushi <imtsukushi@ya.ru>"]
readme = "README.md"
packages = [
    {include = "gendiff"}
]
classifiers = [
        "Development Status :: 3 - Alpha",
        "Environment :: Console",
        "Framework :: Flake8",
        "License :: OSI Approved :: MIT License",
        "Natural Language :: Russian",
        "Operating System :: POSIX :: Linux",
        "Programming Language :: Python :: 3.10",
	]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
